MT#59962 xmlrpc2di: int is long indeed
Properly select the equal operator based on long instead.

Otherwise it selects the `ValueStruct` as the latest
possible from the variant based `_value` and eventually
it gets stumbled over `[]` operator for
building structs which asserts it.

+++ apps/xmlrpc2di/XMLRPC2DI.cpp
@@ -642,7 +642,7 @@ void XMLRPC2DIServer::xmlrpcval2amargarray(XmlRpcValue& v, AmArg& a,
 void XMLRPC2DIServer::xmlrpcval2amarg(XmlRpcValue& v, AmArg& a) {
   if (v.valid()) {
     switch (v.getType()) {
-    case XmlRpcValue::TypeInt:   {  /* DBG("X->A INT\n"); */ a = (int)v;    }  break;
+    case XmlRpcValue::TypeInt:   {  /* DBG("X->A INT\n"); */ a = (long)v;    }  break;
     case XmlRpcValue::TypeDouble:{  /*  DBG("X->A DBL\n"); */ a = (double)v; }  break;
     case XmlRpcValue::TypeString:{  /*  DBG("X->A STR\n"); */ a = ((string)v).c_str(); }  break;
     case XmlRpcValue::TypeBoolean : { /*   DBG("X->A BOL\n"); */ a = (bool)v;  } break;
